requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
openai==1.3.0
python-multipart==0.0.6
pydantic==2.5.0
redis==5.0.1
asyncio-mqtt==0.13.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
numpy==1.24.3
pyttsx3==2.90
pyaudio==0.2.11 